requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
boto3
python-dotenv
